refactor(cli): split pipeline.rs into a pipeline/ submodule (#193)
At ~840 lines pipeline.rs was over the project's ~500-line split
heuristic. Cut along the natural seam with no behavior change:

- pipeline/outcomes.rs — the three outcome enums (RunOutcome,
  SyncPassOutcome, InspectOutcome) + RunOutcome::advanced_state + their
  pure variant-level tests.
- pipeline/passes.rs — the four sync passes + gather_copy_clocks +
  silent_merge_clock + the LUB-contract tests.
- pipeline/mod.rs — the seam: module docs + `pub use` re-exports, so the
  public API (`pipeline::{run_one, RunOutcome, sync_pass_*, ...}`) is
  byte-for-byte unchanged for daemon.rs / main.rs / the integration tests.

Verified: cli lib (154) + all integration suites green; clippy -D
warnings clean; `cargo doc` introduces no new broken intra-doc links
(14 pre-existing on main, 14 after — unchanged).
